+++ src/pathological/files.clj
@@ -8,11 +8,11 @@
     [pathological.utils :as u])
   (:import
     [java.nio.file DirectoryStream$Filter
-     Files
-     FileVisitor
-     Path]
+                   Files
+                   FileVisitor
+                   Path FileAlreadyExistsException]
     [java.nio.charset Charset
-     StandardCharsets]
+                      StandardCharsets]
     [java.util.function BiPredicate]
     [java.io InputStream OutputStream]))
 
@@ -89,12 +89,12 @@
 
 (defn write-lines
   ([^Path path ^Iterable lines]
-    (write-lines path lines :utf-8))
+   (write-lines path lines :utf-8))
   ([^Path path ^Iterable lines charset & options]
-    (let [^Charset charset (u/->charset charset)
-          ^"[Ljava.nio.file.OpenOption;"
-          open-options (u/->open-options-array options)]
-      (Files/write path lines charset open-options))))
+   (let [^Charset charset (u/->charset charset)
+         ^"[Ljava.nio.file.OpenOption;"
+         open-options (u/->open-options-array options)]
+     (Files/write path lines charset open-options))))
 
 (defn read-all-bytes
   [^Path path]
@@ -102,24 +102,24 @@
 
 (defn read-all-lines
   ([^Path path]
-    (read-all-lines path StandardCharsets/UTF_8))
+   (read-all-lines path StandardCharsets/UTF_8))
   ([^Path path ^Charset charset]
-    (Files/readAllLines path (u/->charset charset))))
+   (Files/readAllLines path (u/->charset charset))))
 
 (defn lines-stream
   ([^Path path]
-    (Files/lines path))
+   (Files/lines path))
   ([^Path path charset]
-    (Files/lines path (u/->charset charset))))
+   (Files/lines path (u/->charset charset))))
 
 (defn lines
   ([^Path path]
-    (u/stream-seq (lines-stream path)))
+   (u/stream-seq (lines-stream path)))
   ([^Path path charset]
-    (u/stream-seq (lines-stream path charset))))
+   (u/stream-seq (lines-stream path charset))))
 
 (deftype FnBackedBiPredicate
-         [predicate-fn]
+  [predicate-fn]
 
   BiPredicate
   (test [_ path basic-file-attributes]
@@ -314,7 +314,7 @@
   (Files/isExecutable path))
 
 (deftype FnBackedDirectoryStreamFilter
-         [filter-fn]
+  [filter-fn]
 
   DirectoryStream$Filter
   (accept [_ path]
@@ -322,13 +322,13 @@
 
 (defn new-directory-stream
   ([^Path path]
-    (Files/newDirectoryStream path))
+   (Files/newDirectoryStream path))
   ([^Path path glob-or-filter]
-    (if (instance? String glob-or-filter)
-      (Files/newDirectoryStream path ^String glob-or-filter)
-      (Files/newDirectoryStream path
-        ^DirectoryStream$Filter
-        (->FnBackedDirectoryStreamFilter glob-or-filter)))))
+   (if (instance? String glob-or-filter)
+     (Files/newDirectoryStream path ^String glob-or-filter)
+     (Files/newDirectoryStream path
+       ^DirectoryStream$Filter
+       (->FnBackedDirectoryStreamFilter glob-or-filter)))))
 
 (defn new-input-stream
   [^Path path & options]
@@ -345,16 +345,16 @@
 (defn new-buffered-reader
   ([^Path path] (new-buffered-reader path :utf-8))
   ([^Path path charset]
-    (Files/newBufferedReader path (u/->charset charset))))
+   (Files/newBufferedReader path (u/->charset charset))))
 
 (defn new-buffered-writer
   ([^Path path & args]
-    (let [[first & rest] args
-          charset (u/->charset (if (u/charset? first) first :utf-8))
-          ^"[Ljava.nio.file.OpenOption;"
-          open-options (u/->open-options-array
-                         (if (u/charset? first) rest args))]
-      (Files/newBufferedWriter path charset open-options))))
+   (let [[first & rest] args
+         charset (u/->charset (if (u/charset? first) first :utf-8))
+         ^"[Ljava.nio.file.OpenOption;"
+         open-options (u/->open-options-array
+                        (if (u/charset? first) rest args))]
+     (Files/newBufferedWriter path charset open-options))))
 
 (defn walk-stream
   [^Path path
@@ -380,11 +380,11 @@
     (u/->file-visit-result control)))
 
 (deftype FnBackedFileVisitor
-         [pre-visit-directory-fn
-          post-visit-directory-fn
-          visit-file-fn
-          visit-file-failed-fn
-          accumulator-atom]
+  [pre-visit-directory-fn
+   post-visit-directory-fn
+   visit-file-fn
+   visit-file-failed-fn
+   accumulator-atom]
 
   FileVisitor
   (preVisitDirectory [_ directory basic-file-attributes]
@@ -439,37 +439,71 @@
     (map? first) [(assoc first :type :file) second]))
 
 (defn populate-file-tree
-  [^Path path definition]
-  (doseq [[name & rest] definition
-          :let [[attributes rest] (parse-definition rest)
-                path (p/path path (clojure.core/name name))]]
-    (condp = (:type attributes)
-      :directory
-      (do
-        (create-directories path)
-        (when (seq rest)
-          (populate-file-tree path rest)))
+  [^Path path definition & {:as options}]
+  (let [{:keys [on-exists]
+         :or   {on-exists :throw}} options]
+    (doseq [[name & rest] definition
+            :let [[attributes rest] (parse-definition rest)
+                  path (p/path path (clojure.core/name name))]]
+      (condp = (:type attributes)
+        :directory
+        (do
+          (create-directories path)
+          (when (seq rest)
+            (populate-file-tree path rest)))
 
-      :file
-      (do
-        (create-file path)
-        (when (:content attributes)
-          (write-lines path (:content attributes))))
+        :file
+        (let [open-options
+              (cond
+                (= :throw on-exists) #{:create-new :write}
+                (= :skip on-exists) #{:create-new :write}
+                (= :overwrite on-exists) #{:create :truncate-existing :write})
+              content (get attributes :content [])]
+          (try
+            (apply write-lines path content :utf-8 open-options)
+            (catch FileAlreadyExistsException exception
+              (when-not (= :skip on-exists)
+                (throw exception)))))
 
-      :symbolic-link
-      (do
-        (assert (:target attributes)
-          (str "Attribute :target missing for path " path))
-        (create-symbolic-link
-          path
-          (p/path (p/file-system path) (:target attributes))))
+        :symbolic-link
+        (let [target (:target attributes)
+              create-fn
+              (fn [path target]
+                (create-symbolic-link path
+                  (p/path (p/file-system path) target)))]
+          (assert target (str "Attribute :target missing for path: " path))
+          (try
+            (create-fn path target)
+            (catch FileAlreadyExistsException exception
+              (cond
+                (= :overwrite on-exists)
+                (do
+                  (delete path)
+                  (create-fn path target))
 
-      :link
-      (do
-        (assert (:target attributes)
-          (str "Attribute :target missing for path " path))
-        (create-link path
-          (p/path (p/file-system path) (:target attributes)))))))
+                (= :skip on-exists) nil
+
+                :else (throw exception)))))
+
+        :link
+        (let [target (:target attributes)
+              create-fn
+              (fn [path target]
+                (create-link path
+                  (p/path (p/file-system path) target)))]
+          (assert target (str "Attribute :target missing for path: " path))
+          (try
+            (create-fn path target)
+            (catch FileAlreadyExistsException exception
+              (cond
+                (= :overwrite on-exists)
+                (do
+                  (delete path)
+                  (create-fn path target))
+
+                (= :skip on-exists) nil
+
+                :else (throw exception)))))))))
 
 (defn delete-recursively
   [^Path path]